Create acceptance test OccAppManagementContext
