+++ tests/acceptance/config/behat.yml
@@ -134,6 +134,7 @@ default:
         - FeatureContext: *common_feature_context_params
         - EmailContext:
         - OccContext:
+        - OccAppManagementContext:
         - OccUsersGroupsContext:
 
     cliMain:
